Replaceable output generator function
