+++ index.js
@@ -1,11 +1,8 @@
 var merge = require('lodash.merge')
 
-var getAssetKind = require('./lib/getAssetKind')
-var isHMRUpdate = require('./lib/isHMRUpdate')
-var isSourceMap = require('./lib/isSourceMap')
-
 var createQueuedWriter = require('./lib/output/createQueuedWriter')
 var createOutputWriter = require('./lib/output/createOutputWriter')
+var generateAssetsManifest = require('./lib/generateAssetsManifest')
 
 function AssetsWebpackPlugin (options) {
   this.options = merge({}, {
@@ -14,7 +11,8 @@ function AssetsWebpackPlugin (options) {
     prettyPrint: false,
     update: false,
     fullPath: true,
-    assetsRegex: /\.(jpe?g|png|gif|svg)$/
+    assetsRegex: /\.(jpe?g|png|gif|svg)$/,
+    generator: generateAssetsManifest
   }, options)
   this.writer = createQueuedWriter(createOutputWriter(this.options))
 }
@@ -27,8 +25,7 @@ AssetsWebpackPlugin.prototype = {
     var self = this
 
     compiler.plugin('after-emit', function (compilation, callback) {
-      var output = {}
-      var options = compiler.options
+      var compilerOptions = compiler.options;
       var stats = compilation.getStats().toJson({
         hash: true,
         publicPath: true,
@@ -39,47 +36,8 @@ AssetsWebpackPlugin.prototype = {
         errorDetails: false,
         timings: false
       })
-            // publicPath with resolved [hash] placeholder
 
-      var assetPath = (stats.publicPath && self.options.fullPath) ? stats.publicPath : ''
-            // assetsByChunkName contains a hash with the bundle names and the produced files
-            // e.g. { one: 'one-bundle.js', two: 'two-bundle.js' }
-            // in some cases (when using a plugin or source maps) it might contain an array of produced files
-            // e.g. {
-            // main:
-            //   [ 'index-bundle-42b6e1ec4fa8c5f0303e.js',
-            //     'index-bundle-42b6e1ec4fa8c5f0303e.js.map' ]
-            // }
-      var assetsByChunkName = stats.assetsByChunkName
-
-      output.entries = Object.keys(assetsByChunkName).reduce(function (chunkMap, chunkName) {
-        var assets = assetsByChunkName[chunkName]
-        if (!Array.isArray(assets)) {
-          assets = [assets]
-        }
-        chunkMap[chunkName] = assets.reduce(function (typeMap, asset) {
-          if (isHMRUpdate(options, asset) || isSourceMap(options, asset)) {
-            return typeMap
-          }
-
-          var typeName = getAssetKind(options, asset)
-          typeMap[typeName] = assetPath + asset
-
-          return typeMap
-        }, {})
-
-        return chunkMap
-      }, {})
-
-      output.assets = stats.assets.filter(function (asset) {
-        return self.options.assetsRegex.test(asset.name)
-      }).map(function (asset) {
-        return { name: asset.name, path: assetPath + asset.name }
-      })
-
-      if (self.options.metadata) {
-        output.metadata = self.options.metadata
-      }
+      var output = self.options.generator(stats, self.options, compilerOptions);
 
       self.writer(output, function (err) {
         if (err) {
